23.07.29
유물 Data 추가

코드리팩토링
플레이어

+++ project cosmic/Assets/Script/003LandScn/enemy/EnemyCon.cs
@@ -20,7 +20,7 @@ public class EnemyCon : MonoBehaviour
 
     public int mobNum;
 
-    public int hp;
+    public float hp;
     public int damage;
     public float moveSpeed;
     public float attackDelay;
